+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 1.9.2
 * Removed the movement impairing descriptor from staggered
+* Fixed issue with DR config not working correctly
 
 ## Version 1.9.1b
 * Fixed issue with Aeon Limitless gaze crashing
