+++ TabletopTweaks/Bugfixes/Features/Conditions.cs
@@ -1,22 +1,9 @@
 ﻿using HarmonyLib;
 using Kingmaker.Blueprints;
-using Kingmaker.Blueprints.Classes;
-using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Blueprints.JsonSystem;
-using Kingmaker.EntitySystem;
-using Kingmaker.EntitySystem.Stats;
-using Kingmaker.Enums;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
-using Kingmaker.UnitLogic.FactLogic;
-using Kingmaker.Utility;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Reflection.Emit;
 using TabletopTweaks.Config;
-using TabletopTweaks.Extensions;
 
 namespace TabletopTweaks.Bugfixes.Features {
     class Conditions {
@@ -30,7 +17,7 @@ namespace TabletopTweaks.Bugfixes.Features {
                 Main.LogHeader("Patching Conditions");
                 PatchStaggered();
 
-                static void PatchStaggered(){
+                static void PatchStaggered() {
                     if (ModSettings.Fixes.BaseFixes.IsDisabled("StaggeredDescriptors")) { return; }
                     var Staggered = Resources.GetBlueprint<BlueprintBuff>("df3950af5a783bd4d91ab73eb8fa0fd3");
                     Staggered.GetComponent<SpellDescriptorComponent>().Descriptor = SpellDescriptor.Staggered;
