+++ TabletopTweaks/Config/Homebrew.cs
@@ -16,7 +16,7 @@ namespace TabletopTweaks.Config {
     public class MythicReworkGroup : IDisableableGroup, ICollapseableGroup {
         public bool IsExpanded = true;
         ref bool ICollapseableGroup.IsExpanded() => ref IsExpanded;
-        public void SetExpanded(bool value)  => IsExpanded = value;
+        public void SetExpanded(bool value) => IsExpanded = value;
         public bool DisableAll = false;
         public bool GroupIsDisabled() => DisableAll;
         public void SetGroupDisabled(bool value) => DisableAll = value;
