+++ TabletopTweaks/Config/ICollapseableGroup.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TabletopTweaks.Config {
+﻿namespace TabletopTweaks.Config {
     public interface ICollapseableGroup {
         ref bool IsExpanded();
         void SetExpanded(bool value);
