+++ TabletopTweaks/NewComponents/AbilitySpecific/QuickChannelComponent.cs
@@ -27,7 +27,7 @@ namespace TabletopTweaks.NewComponents.AbilitySpecific {
                         m_IsSpendResource = ability.ResourceLogic.IsSpendResource,
                         Amount = ability.ResourceLogic.CalculateCost(ability) * 2,
                     },
-                    
+
                     CustomActionType = UnitCommand.CommandType.Move
                 };
                 AbilityData.AddAbilityUnique(ref conversionList, moveAbility);
