+++ TabletopTweaks/NewComponents/NestedPsudoActivatableAbilities.cs
@@ -8,7 +8,6 @@ using Kingmaker.PubSubSystem;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Abilities;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
-using Kingmaker.UnitLogic.Abilities.Components;
 using System;
 using System.Collections.Generic;
 using System.Linq;
