+++ TabletopTweaks/NewContent/FighterAdvancedWeaponTrainings/FocusedWeapon.cs
@@ -23,6 +23,9 @@ namespace TabletopTweaks.NewContent.FighterAdvancedWeaponTrainings {
                     " level as his warpriest level. The fighter must have Weapon Focus and Weapon Training with the selected weapon in order to choose this option.\n" +
                     "A focused weapon deals 1d6 damage, this increases to 1d8 at level 6, 1d10 at level 10, 2d6 at level 15, and 2d8 at level 20");
                 bp.m_Icon = WeaponFinesse.Icon;
+                bp.AddComponent<FocusedWeaponDamageComponent>(c => {
+                    c.CheckedClass = FighterClass.ToReference<BlueprintCharacterClassReference>();
+                });
             });
             var FocusedWeaponToggleAbility = Helpers.CreateBlueprint<BlueprintActivatableAbility>("FocusedWeaponToggleAbility", bp => {
                 bp.SetName("Focused Weapon");
@@ -41,10 +44,6 @@ namespace TabletopTweaks.NewContent.FighterAdvancedWeaponTrainings {
                 bp.SetDescription(FocusedWeaponBuff.Description);
                 bp.ParameterType = FeatureParameterType.WeaponCategory;
                 bp.m_Prerequisite = WeaponFocus.ToReference<BlueprintParametrizedFeatureReference>();
-                bp.AddComponent(Helpers.Create<FocusedWeaponComponent>(c => {
-                    c.CheckedClass = FighterClass.ToReference<BlueprintCharacterClassReference>();
-                    c.ToggleBuff = FocusedWeaponBuff.ToReference<BlueprintBuffReference>();
-                }));
                 bp.AddComponent(Helpers.Create<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] {
                         FocusedWeaponToggleAbility.ToReference<BlueprintUnitFactReference>(),
@@ -54,6 +53,7 @@ namespace TabletopTweaks.NewContent.FighterAdvancedWeaponTrainings {
                     c.m_CharacterClass = FighterClass.ToReference<BlueprintCharacterClassReference>();
                     c.Level = 9;
                 }));
+                bp.AddComponent<FocusedWeaponComponent>();
             });
 
             if (ModSettings.AddedContent.FighterAdvancedWeaponTraining.IsDisabled("FocusedWeapon")) { return; }
