+++ TabletopTweaks/NewUI/MechanicActionBarSlotPseudoActivatableAbility.cs
@@ -8,8 +8,8 @@ namespace TabletopTweaks.NewUI {
 
         public BlueprintBuffReference BuffToWatch;
 
-        public override bool IsActive() { 
-            return Unit.Descriptor.HasFact(BuffToWatch); 
+        public override bool IsActive() {
+            return Unit.Descriptor.HasFact(BuffToWatch);
         }
 
         public override int GetResource() {
